+++ tests/testthat/test-funs_string-lazy.R
@@ -18,7 +18,7 @@ test_that("case functions work", {
   )
   test <- as_polars_lf(test_df)
 
-  for (i in c("toupper", "tolower", "str_to_lower", "str_to_upper", "nchar")) {
+  for (i in c("toupper", "tolower", "str_to_lower", "str_to_upper")) {
     pol <- paste0("mutate(test, foo = ", i, "(x1))") |>
       str2lang() |>
       eval() |>
@@ -354,6 +354,25 @@ test_that("length functions work", {
       pull(foo),
     mutate(test_df, foo = str_length(x4)) |>
       pull(foo)
+  )
+
+  expect_equal_lazy(
+    mutate(test, foo = nchar(x4)) |>
+      pull(foo),
+    mutate(test_df, foo = nchar(x4)) |>
+      pull(foo)
+  )
+
+  expect_equal_lazy(
+    mutate(test, foo = nchar(x4, "bytes")) |>
+      pull(foo),
+    mutate(test_df, foo = nchar(x4, "bytes")) |>
+      pull(foo)
+  )
+
+  expect_snapshot_lazy(
+    mutate(test, foo = nchar(x4, "foo")),
+    error = TRUE
   )
 })